% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateRP.R
\name{generateRP}
\alias{generateRP}
\title{Generation of random matrices}
\usage{
generateRP(p, d, B)
}
\arguments{
\item{p}{The original number of variables.}

\item{d}{The reduced dimension.}

\item{B}{The number of projections.}
}
\value{
A single matrix of dimension \code{p} by \code{d*B} containing \code{B} random matrices of dimension \code{p} by \code{d}.
}
\description{
This function generates \code{B} random matrices of dimension \code{p} by \code{d} by using the Haar measure.
}
\examples{
R<-generateRP(p=100,d=2,B=10)
dim(R)
}
